refactor(namedpipe): renamed named pipe behavour & move to core lib
